fixed naming issues
