+++ frontend/src/components/CheckInComponent.jsx
